@@ -1,10 +1,25 @@
 {
-  "name": "seamless-auth-nextjs",
+  "name": "@seamless-auth/nextjs",
   "version": "0.0.0",
   "description": "\"A drop in authentication solution for nextjs applications.",
   "homepage": "https://github.com/fells-code/seamless-auth-nextjs#readme",
   "bugs": {
     "url": "https://github.com/fells-code/seamless-auth-nextjs/issues"
+  },
+  "type": "module",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist",
+    "package.json",
+    "README.md"
+  ],
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "require": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
   },
   "repository": {
     "type": "git",
@@ -12,28 +27,18 @@
   },
   "license": "MIT",
   "author": "Brandon Corbett <brandon.corbett@fellscode.com>",
-  "type": "module",
-  "main": "index.ts",
   "scripts": {
     "test": "jest",
-    "build": "tsc && rollup -c",
-    "dev": "tsc --watch",
-    "prepublishOnly": "npm run build"
+    "build": "rm -rf dist && tsc && rollup -c",
+    "prepare": "npm run build"
   },
   "peerDependencies": {
     "next": ">=15.0.0",
-    "react": "^18.0.0 || ^19.0.0"
+    "react": "^19.1.0"
   },
   "devDependencies": {
     "@types/react": "^19.1.8",
-    "@babel/cli": "^7.24.8",
-    "@babel/core": "^7.25.2",
-    "@babel/plugin-transform-runtime": "^7.27.1",
-    "@babel/preset-env": "^7.25.4",
-    "@babel/preset-react": "^7.24.7",
-    "@babel/preset-typescript": "^7.27.1",
     "@eslint/js": "^9.17.0",
-    "@rollup/plugin-babel": "^6.0.4",
     "@rollup/plugin-commonjs": "^26.0.1",
     "@rollup/plugin-node-resolve": "^15.2.3",
     "@rollup/plugin-terser": "^0.4.4",
@@ -46,7 +51,7 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.0.1",
     "@types/jest": "^29.5.12",
-    "@types/react-dom": "^18.3.0",
+    "@types/react-dom": "^19.1.0",
     "@types/validator": "^13.15.0",
     "eslint": "^9.17.0",
     "eslint-plugin-import": "^2.31.0",
@@ -61,8 +66,7 @@
     "postcss-import": "^16.1.0",
     "postcss-modules": "^6.0.1",
     "postcss-url": "^10.1.3",
-    "react": "^18.3.1",
-    "rimraf": "^6.0.1",
+    "react": "^19",
     "rollup": "^4.0.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
