@@ -6,9 +6,9 @@ import {
 import { useAuth } from "hooks/AuthProvider";
 import { useInternalAuth } from "context/InternalAuthContext";
 import React, { useEffect, useState } from "react";
-import { useRouter } from "next/router";
+import { useRouter } from 'next/navigation'
 
-import styles from "./styles/registerPasskey.module.css";
+import styles from "../styles/registerPasskey.module.css";
 import { isPasskeySupported } from "../utils";
 
 const RegisterPasskey: React.FC = () => {
