@@ -1,4 +1,3 @@
-import babel from "@rollup/plugin-babel";
 import commonjs from "@rollup/plugin-commonjs";
 import resolve from "@rollup/plugin-node-resolve";
 import terser from "@rollup/plugin-terser";
@@ -11,7 +10,7 @@ export default [
     input: "src/index.ts",
     output: {
       dir: "dist",
-      entryFileNames: "[name].[format].js",
+      entryFileNames: "index.js",
       format: "esm",
       sourcemap: true,
     },
@@ -30,13 +29,8 @@ export default [
         inject: true,
         minimize: true,
       }),
-      babel({
-        babelHelpers: "bundled",
-        extensions: [".ts", ".tsx", ".js", ".jsx"],
-        presets: ["@babel/preset-env", "@babel/preset-react"],
-        exclude: "node_modules/**",
-      }),
       terser(),
     ],
+    external: ["react", "react-dom"],
   },
 ];
